pyproject.toml
[project]
name = "rpgate-tg"
version = "0.1.0"
description = "AI-powered Game Master Telegram Bot"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiogram==3.13.0",
    "asyncpg>=0.30.0",
    "fastapi>=0.121.0",
    "httpx>=0.28.1",
    "openai>=2.7.1",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
package = true

[project.scripts]
start = "app.main:main"

[dependency-groups]
dev = [
    "black>=25.9.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "ruff>=0.14.3",
]
